code cleanup + a bunch of other stuff

+++ src/main/java/com/golem/skyblockutils/models/gui/GuiPosition.java
@@ -4,7 +4,6 @@ package com.golem.skyblockutils.models.gui;
 import com.google.gson.JsonObject;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 public class GuiPosition implements Serializable {
 	private double x;
@@ -22,24 +21,24 @@ public class GuiPosition implements Serializable {
 		this.scale = data.get("scale").getAsDouble();
 	}
 
-	public double getX() {
-		return x;
+	public int getX() {
+		return (int) x;
 	}
 
 	public void setX(double x) {
 		this.x = x;
 	}
 
-	public double getY() {
-		return y;
+	public int getY() {
+		return (int) y;
 	}
 
 	public void setY(double y) {
 		this.y = y;
 	}
 
-	public double getScale() {
-		return scale;
+	public float getScale() {
+		return (float) scale;
 	}
 
 	public void setScale(double scale) {
